Move mvi vm to separate package
