+++ app/src/main/java/com/futurice/freesound/arch/mvi/viewmodel/UglyViewModelProviderBridge.kt
@@ -14,15 +14,17 @@
  * limitations under the License.
  */
 
-package com.futurice.freesound.arch.mvi
+package com.futurice.freesound.arch.mvi.viewmodel
 
-import android.arch.lifecycle.LiveData
-import android.arch.lifecycle.ViewModel
+import android.support.v4.app.Fragment
+import com.futurice.freesound.arch.mvi.Action
+import com.futurice.freesound.arch.mvi.Event
+import com.futurice.freesound.arch.mvi.Result
+import com.futurice.freesound.arch.mvi.State
 
-abstract class ViewModel<in E, M> : ViewModel() {
-
-    abstract fun uiEvents(uiEvent: E)
-
-    abstract fun uiModels(): LiveData<M>
-
+// Need this bridge for now because you can't call reified functions from Java.
+internal fun
+        <E : Event, A : Action, R : Result, S : State>
+        Fragment.createViewModel(provider: () -> BaseViewModel<E, A, R, S>): BaseViewModel<E, A, R, S> {
+    return viewModelProvider(provider)
 }
