+++ app/src/main/java/com/futurice/freesound/arch/mvi/viewmodel/ViewModel.kt
@@ -14,13 +14,15 @@
  * limitations under the License.
  */
 
-package com.futurice.freesound.arch.mvi
+package com.futurice.freesound.arch.mvi.viewmodel
 
-import android.support.v4.app.Fragment
+import android.arch.lifecycle.LiveData
+import android.arch.lifecycle.ViewModel
 
-// Need this bridge for now because you can't call reified functions from Java.
-internal fun
-        <E : Event, A : Action, R : Result, S : State>
-        Fragment.createViewModel(provider: () -> BaseViewModel<E, A, R, S>): BaseViewModel<E, A, R, S> {
-    return viewModelProvider(provider)
+abstract class ViewModel<in E, M> : ViewModel() {
+
+    abstract fun uiEvents(uiEvent: E)
+
+    abstract fun uiModels(): LiveData<M>
+
 }
